correction for extruders
My printer was always blocking at a certain point. The extrusion of filament stopped and the filament was pressed into the tubes, which causes the unability to remove the filament.
Extra problem is that the models had too mutch of wires and elephant feets.

Solution: the steps/mm for the extruders was too high